+++ src/badge/Badge.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import Badge from 'badge'
+import Badge from './Badge'
 import { CheckIcon } from 'icon/checkIcon'
 
 it('Should read the aria-label attribute and not the children if both are provided', () => {
