+++ src/badge/story.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 
 import { storiesOf } from '@storybook/react'
 import { withKnobs, text } from '@storybook/addon-knobs'
-
-import Badge from 'badge'
+import Badge from './index'
 
 const stories = storiesOf('Badge', module)
 stories.addDecorator(withKnobs)
